Get allowed access tokens from environment
- The easiest safe place to store these tokens at the moment is in the
  environment.
- Factor out the two test helper functions into their own module as
  they were starting to dominate the test module.
