--- config/tokens
@@ -1,1 +1,0 @@
-valid-token
